Refactor / Document metrics module
